requirements.txt
torch==2.2.0
torchtext==0.16.2
pandas==1.5.3
scikit-learn==1.2.2
matplotlib==3.7.1
seaborn==0.12.2
tqdm==4.65.0
PyYAML==6.0.2